Create background.js
n this code, the background script listens for the click event on the browser action button (the extension button in the Chrome toolbar). When the button is clicked, it sends a message to the content script to perform the action of marking clickbait. The chrome.tabs.sendMessage function is used to send the message to the content script in the active tab.

Remember to include this code in your project's background.js file and make sure it is properly loaded in the manifest.json file.